refactor: Parser 객체 분리 및 PurchaseAmount 유효성 검증
