+++ docs/README.md
@@ -1,11 +1,10 @@
-
-
 # 🛠기능 구현 목록
 
-- [ ] __구입 금액 입력 받기__
-    - [ ] 1000원 단위의 금액 외 입력시 예외처리
-    - [ ] null값 입력시 예외처리
-    - [ ] 정수 외값 입력시 예외처리
+- [x] __구입 금액 입력 받기__
+    - [x] 구입 입력 요청 메세지 출력
+    - [x] 1000원 단위의 금액 외 입력시 예외처리
+    - [x] null값 입력시 예외처리
+    - [x] 정수 외값 입력시 예외처리
 
 - [ ] __구입한 로또 갯수 계산하기 => 한 장당 1000원으로 계산 ex)8000원 입력시 8장 구매__
     - [ ] 로또 구입 갯수 메세지 출력
@@ -44,7 +43,6 @@
 - [ ] __총 수익률 계산__
     - [ ] 로또 당첨금 / 로또 구입비 백분율로 계산
     - [ ] 둘째 자리에서 반올림
-
 
 #### 실행 결과 예시
 >
