--- src/main/java/lotto/controller.java
@@ -1,2 +1,0 @@
-package lotto;public class controller {
-}
