--- src/main/java/lotto/domain/LottoMachine.java
@@ -1,2 +1,0 @@
-package lotto.domain;public class LottoMachine {
-}
