+++ src/main/java/lotto/utils/Parser.java
@@ -1,2 +1,8 @@
-package lotto.utils;public class Parser {
+package lotto.utils;
+
+public class Parser {
+    public static int parseStringToInt(String stringValue) {
+            return Integer.parseInt(stringValue);
+    }
+
 }
