+++ src/main/java/lotto/view/OutputView.java
@@ -1,4 +1,4 @@
-package view;
+package lotto.view;
 
 public class OutputView {
 }
